+++ pytest_play/engine.py
@@ -174,17 +174,6 @@ class PlayEngine(object):
             elapsed = time.time() - start_time
             print(dict(command, _elapsed=elapsed))
             self.update_variables({'_elapsed': elapsed})
-            if 'property_name' in command:
-                self.update_variables({command['property_name']: elapsed})
-                if self.request.config.getoption('--junit-xml'):
-                    # if property_name key in commands, track elapsed time
-                    # (e.g., time needed from the previous click on the login
-                    # button until you are able to interact successfully
-                    # with the application). This way you can track
-                    # this time so that will be included in a machine
-                    # interpretable report if --junit-xml cli option
-                    # has been used
-                    self.record_property(command['property_name'], elapsed)
         return return_value
 
     def update_variables(self, extra_variables):