+++ pytest_play/providers/metrics.py
@@ -7,6 +7,17 @@ class MetricsProvider(BaseProvider):
     def command_record_property(self, command, **kwargs):
         """ Metrics scenario """
         name = command['name']
+        expression = command['expression']
+        value = self.engine.execute_command(
+            {'provider': 'python',
+             'type': 'exec',
+             'expression': expression})
+        self.engine.update_variables({name: value})
+        self.engine.record_property(name, value)
+
+    def command_record_elapsed(self, command, **kwargs):
+        """ Metrics scenario """
+        name = command['name']
         value = self.engine.variables['_elapsed']
         self.engine.update_variables({name: value})
         self.engine.record_property(name, value)